move state generation
